+++ dispatches/models/fossil_case/ultra_supercritical_plant/storage/discharge_design_ultra_supercritical_power_plant.py
@@ -71,7 +71,7 @@ from dispatches.models.fossil_case.ultra_supercritical_plant import (
     ultra_supercritical_powerplant as usc)
 
 # Import properties package for Solar salt
-from dispatches.models.fossil_case.properties import solarsalt_properties
+from dispatches.properties import solarsalt_properties
 
 
 scaling_obj = 1e-7
