+++ dispatches/models/fossil_case/ultra_supercritical_plant/storage/tests/test_charge_usc_powerplant.py
@@ -32,7 +32,8 @@ from idaes.core.util import get_solver
 
 from dispatches.models.fossil_case.ultra_supercritical_plant import (
     ultra_supercritical_powerplant as usc)
-from . import charge_design_ultra_supercritical_power_plant as charge_usc
+from dispatches.models.fossil_case.ultra_supercritical_plant.storage \
+     import charge_design_ultra_supercritical_power_plant as charge_usc
 
 
 optarg = {"max_iter": 300,
