+++ dispatches/models/fossil_case/ultra_supercritical_plant/storage/tests/test_discharge_usc_powerplant.py
@@ -32,7 +32,8 @@ from idaes.core.util import get_solver
 from dispatches.models.fossil_case.ultra_supercritical_plant import (
     ultra_supercritical_powerplant as usc)
 
-from . import discharge_design_ultra_supercritical_power_plant as discharge_usc
+from dispatches.models.fossil_case.ultra_supercritical_plant.storage \
+     import discharge_design_ultra_supercritical_power_plant as discharge_usc
 
 
 optarg = {"max_iter": 300,
