+++ dispatches/models/nuclear_case/lmp_signal.json
