+++ dispatches/models/nuclear_case/sample_lmp_plots.png
