+++ dispatches/models/nuclear_case/schematic_nuclear_flowsheet.png
