+++ dispatches/properties/hitecsalt_properties.py
