+++ dispatches/properties/solarsalt_properties.py
